More reorganizing of files and functions
